sandbox/crs/pyproject.toml
[project]
name = "benchmark"
version = "0.1.0"
dependencies = [
    "aiohttp==3.9.5",
    "aioshutil==1.4",
    "aiorwlock==1.4.0",
    "PyYAML==6.0.1",
    "GitPython==3.1.43",
    "langchain==0.2.7",
    "langchain-community==0.2.7",
    "langchain-openai==0.1.14",
    "langchain-text-splitters==0.2.2",
    "langgraph==0.1.5",
    "psutil==6.0.0",
]

[project.optional-dependencies]
dev = [
    "pip-tools",
    "pyright",
    "black==24.4.2",
    "isort==5.13.2",
    "pre-commit",
    "types-PyYAML",
    "types-requests",
]

[build-system]
requires = ["setuptools", "wheel"]

[tool.pip-tools]
generate-hashes = true

[tool.pyright]
pythonVersion = "3.12"
pythonPlatform = "Linux"
reportMissingTypeStubs = true
exclude = ["typings"]

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 120
include_trailing_comma = true
skip_gitignore = true
known_first_party = ["api", "pipeline", "config", "logger"]

[tool.black]
target-version = ["py312"]
line_length = 120
preview = true
enable-unstable-feature= ["hug_parens_with_braces_and_square_brackets"]
